+++ ruoyi-ui/vite/plugins/index.js
@@ -9,7 +9,7 @@ export default function createVitePlugins(viteEnv, isBuild = false) {
   const vitePlugins = [vue()]
   vitePlugins.push(createAutoImport())
   vitePlugins.push(createSetupExtend())
-  vitePlugins.push(createSvgIcon(isBuild))
+  vitePlugins.push(...createSvgIcon())
   isBuild && vitePlugins.push(...createCompression(viteEnv))
   return vitePlugins
 }