+++ ruoyi-ui/vite/plugins/svg-icon.js
@@ -1,10 +1,23 @@
-import { createSvgIconsPlugin } from 'vite-plugin-svg-icons'
-import path from 'path'
+import VitePluginSvgSpritemap from '@spiriit/vite-plugin-svg-spritemap'
 
-export default function createSvgIcon(isBuild) {
-  return createSvgIconsPlugin({
-    iconDirs: [path.resolve(process.cwd(), 'src/assets/icons/svg')],
-    symbolId: 'icon-[dir]-[name]',
-    svgoOptions: isBuild
+export default function createSvgIcon() {
+  return VitePluginSvgSpritemap('./src/assets/icons/svg/*.svg', {
+    prefix: 'icon-',
+    route: '/__spritemap',
+    oxvg: false,
+    styles: false,
+    injectSvgOnDev: false,
+    output: {
+      filename: '[name].[hash][extname]',
+      name: 'spritemap.svg',
+      view: false,
+      use: false
+    },
+    svgo: {
+      plugins: [
+        { name: 'removeDimensions' },
+        { name: 'removeStyleElement' }
+      ]
+    }
   })
 }